+++ ase/tasks/calcfactory.py
@@ -110,7 +110,7 @@ class CalculatorFactory:
 
 # Recognized names of calculators sorted alphabetically:
 calcnames = ['abinit', 'aims', 'asap', 'castep', 'dftb', 'elk', 'emt',
-             'exciting', 'fleur', 'gpaw', 'hotbit', 'jacapo',
+             'exciting', 'fleur', 'gpaw', 'gaussian', 'hotbit', 'jacapo',
              'lammps', 'lj', 'mopac', 'morse',
              'nwchem', 'siesta', 'turbomole', 'vasp']
 
@@ -118,11 +118,12 @@ classnames = {'asap': 'EMT',
               'elk': 'ELK',
               'emt': 'EMT',
               'fleur': 'FLEUR',
+              'gaussian': 'Gaussian',
               'jacapo': 'Jacapo',
               'lammps': 'LAMMPS',
               'lj': 'LennardJones',
-              'morse': 'MorsePotential',
               'mopac': 'Mopac',
+              'morse': 'MorsePotential',
               'nwchem': 'NWchem',
               'vasp': 'Vasp'}
 
@@ -167,7 +168,7 @@ def calculator_factory(name, **kwargs):
     module = __import__('ase.calculators.' + name, {}, None, [classname])
     Class = getattr(module, classname)
 
-    if name in ['emt', 'lammps', 'lj', 'mopac', 'morse']:
+    if name in ['emt', 'gaussian', 'lammps', 'lj', 'mopac', 'morse']:
         kpts = 'no k-points'
     else:
         kpts = None